Extract `queryToSeries` util
