+++ static/app/views/starfish/utils/queryToSeries.ts
